Daily Report/requirements.txt
Flask==2.3.2
loguru==0.7.0
protobuf==5.26.1
Requests==2.31.0
google-cloud-secret-manager
google-cloud-firestore
functions-framework
